requirements.txt
# This template is a low-dependency template. 
# By default there is no requirements added here.
# Add the requirements you need to this file.
# or run `make init` to create this file automatically based on the template.
# You can also run `make switch-to-poetry` to use the poetry package manager.
typed-argument-parser
boto3
botocore
requests
deepdiff
python-dateutil
types-python-dateutil
pydantic
pyyaml
types-PyYAML
sqlalchemy
unidecode
prettytable
humanize
google-api-python-client
